+++ config/easy/SquareSpiralForAlienContact/code/CGCode.js
@@ -1,0 +1,6 @@
+const crypticInstructions = readline();
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('Square Spiral');